Migrate victory-voronoi to TypeScript
