+++ .changeset/long-files-itch.md
@@ -1,0 +1,5 @@
+---
+"victory-voronoi": patch
+---
+
+Migrate victory-voronoi to TypeScript
