--- packages/victory-voronoi/src/index.js
@@ -1,2 +1,0 @@
-export { default as VictoryVoronoi } from "./victory-voronoi";
-export { default as Voronoi } from "./voronoi";
