+++ packages/victory-voronoi/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-voronoi";
+export * from "./voronoi";
